+++ src/main/java/atwoz/atwoz/member/application/exception/BannedMemberException.java
@@ -1,0 +1,7 @@
+package atwoz.atwoz.member.application.exception;
+
+public class BannedMemberException extends RuntimeException {
+    public BannedMemberException() {
+        super("영구 정지된 유저입니다.");
+    }
+}
