+++ src/main/java/atwoz/atwoz/member/domain/member/exception/InvalidMemberIdException.java
@@ -1,4 +1,4 @@
-package atwoz.atwoz.member.exception;
+package atwoz.atwoz.member.domain.member.exception;
 
 public class InvalidMemberIdException extends RuntimeException {
     public InvalidMemberIdException() {
