--- src/main/java/atwoz/atwoz/member/exception/MemberPermanentStopException.java
@@ -1,7 +1,0 @@
-package atwoz.atwoz.member.exception;
-
-public class MemberPermanentStopException extends RuntimeException {
-    public MemberPermanentStopException() {
-        super("영구 정지된 유저입니다.");
-    }
-}
