+++ src/test/java/atwoz/atwoz/member/MemberProfileUpdateTest.java
@@ -6,10 +6,10 @@ import atwoz.atwoz.job.exception.JobNotFoundException;
 import atwoz.atwoz.member.application.MemberService;
 import atwoz.atwoz.member.application.dto.MemberProfileUpdateRequest;
 import atwoz.atwoz.member.application.dto.MemberProfileUpdateResponse;
+import atwoz.atwoz.member.application.exception.MemberNotFoundException;
 import atwoz.atwoz.member.domain.member.*;
-import atwoz.atwoz.member.exception.InvalidHobbyIdException;
-import atwoz.atwoz.member.exception.InvalidMemberEnumValueException;
-import atwoz.atwoz.member.exception.MemberNotFoundException;
+import atwoz.atwoz.member.domain.member.exception.InvalidHobbyIdException;
+import atwoz.atwoz.member.domain.member.exception.InvalidMemberEnumValueException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
