+++ src/test/java/atwoz/atwoz/member/domain/NickNameTest.java
@@ -1,7 +1,7 @@
 package atwoz.atwoz.member.domain;
 
 import atwoz.atwoz.member.domain.member.Nickname;
-import atwoz.atwoz.member.exception.InvalidNickNameException;
+import atwoz.atwoz.member.domain.member.exception.InvalidNickNameException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
